Delete the src folder and modify its contents
